+++ addon/components/paper-input.js
@@ -52,6 +52,7 @@ export default BaseFocusable.extend(ColorMixin, FlexMixin, ChildMixin, {
    *    true: input has been touched and is invalid.
    */
   isInvalid: computed.or('validationErrorMessages.length', 'isNativeInvalid'),
+  isValid: computed.not('isInvalid'),
 
   isInvalidAndTouched: computed.and('isInvalid', 'isTouched'),
 
@@ -132,8 +133,8 @@ export default BaseFocusable.extend(ColorMixin, FlexMixin, ChildMixin, {
   // Lifecycle hooks
   didReceiveAttrs() {
     this._super(...arguments);
-    assert('{{paper-input}} and {{paper-select}} require an `onChange` action or null for no action.', this.get('onChange') !== undefined);
-    this.notifyInvalid();
+    assert('{{paper-input}} requires an `onChange` action or null for no action.', this.get('onChange') !== undefined);
+    this.notifyValidityChange();
   },
 
   didInsertElement() {
@@ -150,7 +151,7 @@ export default BaseFocusable.extend(ColorMixin, FlexMixin, ChildMixin, {
   },
 
   willClearRender() {
-    this.sendAction('onInvalid', false);
+    this.sendAction('onValidityChange', false);
   },
 
   willDestroyElement() {
@@ -199,12 +200,12 @@ export default BaseFocusable.extend(ColorMixin, FlexMixin, ChildMixin, {
     return offsetHeight + (line > 0 ? line : 0);
   },
 
-  notifyInvalid() {
-    let isInvalid = this.get('isInvalid');
-    let lastIsInvalid = this.get('lastIsInvalid');
-    if (lastIsInvalid !== isInvalid) {
-      this.sendAction('onInvalid', isInvalid);
-      this.set('lastIsInvalid', isInvalid);
+  notifyValidityChange() {
+    let isValid = this.get('isValid');
+    let lastIsValid = this.get('lastIsValid');
+    if (lastIsValid !== isValid) {
+      this.sendAction('onValidityChange', isValid);
+      this.set('lastIsValid', isValid);
     }
   },
 
@@ -220,13 +221,13 @@ export default BaseFocusable.extend(ColorMixin, FlexMixin, ChildMixin, {
       this.growTextarea();
       let inputElement = this.$('input').get(0);
       this.set('isNativeInvalid', inputElement && inputElement.validity && inputElement.validity.badInput);
-      this.notifyInvalid();
+      this.notifyValidityChange();
     },
 
     handleBlur(e) {
       this.sendAction('onBlur', e);
       this.set('isTouched', true);
-      this.notifyInvalid();
+      this.notifyValidityChange();
     }
   }
 });